+++ lib/views/startup-page.dart
@@ -1,11 +1,17 @@
+import 'dart:async';
+
+import 'package:dvote_common/widgets/alerts.dart';
 import 'package:dvote_common/widgets/loading-spinner.dart';
 import 'package:flutter/material.dart';
+import 'package:uni_links/uni_links.dart';
 import 'package:vocdoni/app-config.dart';
 import 'package:vocdoni/lib/analytics.dart';
+import 'package:vocdoni/lib/app-links.dart';
 import 'package:vocdoni/lib/i18n.dart';
 import 'package:vocdoni/lib/logger.dart';
 import 'package:vocdoni/lib/notifications.dart';
 import 'package:vocdoni/lib/startup.dart';
+import 'package:vocdoni/view-modals/action-account-select.dart';
 import '../lib/globals.dart';
 import 'package:dvote_common/widgets/flavor-banner.dart';
 import 'package:vocdoni/lib/extensions.dart';
@@ -41,7 +47,10 @@ class _StartupPageState extends State<StartupPage> {
         .then((_) => Globals.analytics.init())
         .then((_) => Globals.analytics.trackEvent(Events.APP_START))
         .then((_) {
+      // non-blocking start networking
+      startNetworking();
       showNextScreen();
+      initLinking();
 
       // Detached update of the cached bootnodes
       Globals.appState.refresh().catchError(
@@ -60,6 +69,54 @@ class _StartupPageState extends State<StartupPage> {
       // RETRY ITSELF
       Future.delayed(Duration(seconds: 10)).then((_) => initApplication());
     });
+  }
+
+  initLinking() async {
+    // HANDLE APP LAUNCH LINK
+    try {
+      getInitialUri()
+          .then((uri) =>
+              handleLink(uri, context ?? Globals.navigatorKey.currentContext))
+          .catchError((err) => handleIncomingLinkError(err));
+
+      // HANDLE RUNTIME LINKS
+      getUriLinksStream().listen(
+          (uri) =>
+              handleLink(uri, context ?? Globals.navigatorKey.currentContext),
+          onError: handleIncomingLinkError);
+
+      // Display the screen for a notification (if one is pending)
+      Future.delayed(Duration(seconds: 1))
+          .then((_) => Notifications.handlePendingNotification());
+    } catch (err) {
+      logger.log(err);
+      showAlert(getText(context, "main.theLinkYouFollowedAppearsToBeInvalid"),
+          title: getText(context, "main.error"), context: context);
+    }
+  }
+
+  handleLink(Uri givenUri, BuildContext ctx) {
+    if (givenUri == null || !Globals.accountPool.hasValue) return;
+    if (Globals.accountPool.value.length == 1 ||
+        givenUri.path.contains("recovery")) {
+      handleIncomingLink(givenUri, ctx).catchError(handleIncomingLinkError);
+    } else {
+      Navigator.push(Globals.navigatorKey.currentContext,
+              MaterialPageRoute(builder: (context) => LinkAccountSelect()))
+          .then((result) {
+        if (result != null && result is int) {
+          Globals.appState.selectAccount(result);
+          handleIncomingLink(givenUri, ctx).catchError(handleIncomingLinkError);
+        }
+      });
+    }
+  }
+
+  handleIncomingLinkError(err) {
+    logger.log(err?.toString() ?? "handleIncomingLinkError");
+    final ctx = context ?? Globals.navigatorKey.currentContext;
+    showAlert(getText(ctx, "error.thereWasAProblemHandlingTheLink"),
+        title: getText(ctx, "main.error"), context: ctx);
   }
 
   void showNextScreen() {